+++ PetrolPump/ReceiveCreditForm.cs
@@ -89,7 +89,7 @@ namespace PetrolPump
                 TBAmount.BackColor = Color.White;
             }
 
-            if (int.Parse(TBAmount.Text) <= 0)
+            if (double.Parse(TBAmount.Text) <= 0)
             {
                 TBAmount.BackColor = Color.FromArgb(255, 207, 207);
                 return;
